+++ packages/config/src/constants/wrappedTokens.ts
@@ -57,24 +57,11 @@ export const WRAPPED_NATIVE_CURRENCIES: Record<SupportedChainId, TokenInfo> = {
       symbol: 'WAVAX',
     },
   ),
-  [SupportedChainId.LENS]: createWrappedTokenForChain(
-    SupportedChainId.LENS,
-    '0x6bdc36e20d267ff0dd6097799f82e78907105e2f',
-    {
-      decimals: 18,
-      name: 'Wrapped GHO',
-      symbol: 'WGHO',
-    },
-  ),
-  [SupportedChainId.BNB]: createWrappedTokenForChain(
-    SupportedChainId.BNB,
-    '0xbb4CdB9CBd36B01bD1cBaEBF2De08d9173bc095c',
-    {
-      decimals: 18,
-      name: 'Wrapped BNB',
-      symbol: 'WBNB',
-    },
-  ),
+  [SupportedChainId.BNB]: createWrappedTokenForChain(SupportedChainId.BNB, '0xbb4CdB9CBd36B01bD1cBaEBF2De08d9173bc095c', {
+    decimals: 18,
+    name: 'Wrapped BNB',
+    symbol: 'WBNB',
+  }),
   [SupportedChainId.PLASMA]: createWrappedTokenForChain(
     SupportedChainId.PLASMA,
     '0x6100e367285b01f48d07953803a2d8dca5d19873',