+++ packages/subgraph/src/api.spec.ts
@@ -335,7 +335,6 @@ describe('CoW subgraph URL', () => {
       [SupportedChainId.SEPOLIA]: null,
       [SupportedChainId.POLYGON]: null,
       [SupportedChainId.AVALANCHE]: null,
-      [SupportedChainId.LENS]: null,
       [SupportedChainId.BNB]: null,
       [SupportedChainId.LINEA]: null,
       [SupportedChainId.PLASMA]: null,